Updates for PN-706 - add documentation for the CLI
This adds a `relay doc generate` command that can build
man and markdown format doc trees.

The output format is not ideal; I'd like to have a "real"
manpage that has all the subcommands documented in one page,
and same with the markdown. But it's better than nothing.
The output is meant to be checked into the docs repo or viewed
locally.