Reworks the `relay doc` subcommand to work with the website
The command now only generates markdown, not man pages.
It generates a single page that describes _all_ the runnable
subcommands with their flags, like a typical man page
would.

This is both more useful for end-users (the markdown is
pretty readable if you see it on stdout) and the relay.sh
website which wasn't happy about having a dozen tiny files,
one for each subcommand.

The generated doc is also checked into the docs/ subdirectory.